+++ scripts/addTwo.ts
@@ -1,3 +1,3 @@
 export const addTwo = (x: number, y: number): number => {
-  return x - y;
+  return x + y;
 };
